+++ examples/routing_fallback/src/main.rs
@@ -5,7 +5,7 @@ use picoserve::{response::IntoResponse, routing::get};
 struct CustomNotFound;
 
 impl picoserve::routing::PathRouterService<()> for CustomNotFound {
-    async fn call_request_handler_service<
+    async fn call_path_router_service<
         R: picoserve::io::Read,
         W: picoserve::response::ResponseWriter<Error = R::Error>,
     >(