+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Routing types which have `PathParameters` use tuples rather than one of `NoPathParameters`, `OnePathParameter`, `ManyPathParameters`.
 - [`Router`](https://docs.rs/picoserve/latest/picoserve/routing/struct.Router.html) has a different number of type parameters.
 - Updated `picoserve_derive` to version 0.1.4.
+- The method of [`PathRouterService`](https://docs.rs/picoserve/latest/picoserve/routing/trait.PathRouterService.html) is corrected, changing from `call_request_handler_service` to `call_path_router_service`.
 
 ### Fixed
 - The [`Debug`](https://doc.rust-lang.org/core/fmt/trait.Debug.html) implementation of [`HeaderName`](https://docs.rs/picoserve/latest/picoserve/request/struct.HeaderName.html) and [`HeaderValue`](https://docs.rs/picoserve/latest/picoserve/request/struct.HeaderValue.html) includes the surrounding double quotes.